server/pyproject.toml
[project]
name = "openagents-server"
version = "0.1.0"
description = "Multi-agent orchestration server API"
authors = [{ name = "OpenAgents Team" }]
requires-python = ">=3.13"
dependencies = [
    "elevenlabs>=2.23.0",
    "fastapi>=0.104.0",
    "geopy>=2.4.1",
    "google-auth>=2.23.0",
    "googlemaps>=4.10.0",
    "httpx>=0.25.0",
    "livekit-api>=1.1.0",
    "livekit-plugins-turn-detector>=1.3.6",
    "loguru>=0.7.3",
    "mcp[cli]>=1.9.1",
    "openai>=2.8.0",
    "openai-agents>=0.5.1",
    "passlib[bcrypt]>=1.7.4",
    "perplexityai>=0.22.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "pyyaml>=6.0.1",
    "uvicorn[standard]>=0.24.0",
    "wikipedia>=1.4.0",
    "yfinance>=0.2.66",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=server --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
